requirements.txt
Django>=4.1.0,<4.2.0
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<3.0
drf-spectacular>=0.23.1,<0.24
pillow>=9.2.0,<9.3.0
uwsgi>=2.0.19,<2.1